+++ CHANGELOG.md
@@ -7,8 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.1] - 2026-07-15
+## [0.1.1] - 2026-07-15
 
 ### Added
 
 -This changelog
+-Walkthrough context and pop ups
+-Fixed bug in add mission button validation